clean up of code
